+++ packages/shared/src/modules/contributors/libs/types/contributor-get-all-request-dto.type.ts
@@ -1,6 +1,7 @@
 import { type PaginationQueryParameters } from "../../../../libs/types/types.js";
 
 type ContributorGetAllRequestDto = {
+	contributorName?: string | undefined;
 	hasHidden?: boolean | undefined;
 	projectId?: number | undefined;
 } & Partial<PaginationQueryParameters>;